Safari-609.1.11.1/LayoutTests/fast/text/selection-in-initial-advance-region-expected.txt

This tests for a drag select crasher.

لاً
PASS sel.rangeCount is 1
PASS r.startContainer is r.endContainer
PASS r.startOffset is 0
PASS r.endOffset is 3
